+++ CHANGES.rst
@@ -8,6 +8,14 @@ To be released.
 
 - Wheel distributions (``nirum-*.whl``) are now universal between Python 2
   and 3.  [:issue:`78`]
+- ``nirum.rpc.Client`` and its subtype became to raise ``TypeError`` with
+  a better error message when its ``make_request()`` method is overridden and
+  it returns a wrong artity of tuple.  [:issue:`80`]
+- ``nirum.rpc.WsgiApp`` and its subtype became to raise ``TypeError`` with
+  a better error message when its ``make_response()`` method is overridden and
+  it returns a wrong artity of tuple.  [:issue:`80`]
+- Fixed a bug that ``Client.ping()`` method had always raised ``TypeError``.
+  [:issue:`80`]
 
 
 Version 0.5.0